navigation.events: move events to subpackage
