+++ src/main/java/net/kemitix/ldapmanager/navigation/events/NavigationItemUserActionEvent.java
@@ -22,10 +22,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package net.kemitix.ldapmanager.navigation;
+package net.kemitix.ldapmanager.navigation.events;
 
 import lombok.Getter;
 import net.kemitix.ldapmanager.domain.User;
+import net.kemitix.ldapmanager.navigation.UserNavigationItem;
 
 /**
  * Raised when the user presses Enter or Space on a {@link UserNavigationItem}.
