+++ src/main/java/net/kemitix/ldapmanager/navigation/events/NavigationItemUserSelectedEvent.java
@@ -22,10 +22,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package net.kemitix.ldapmanager.navigation;
+package net.kemitix.ldapmanager.navigation.events;
 
 import lombok.Getter;
 import lombok.extern.java.Log;
+import net.kemitix.ldapmanager.navigation.UserNavigationItem;
 
 import javax.naming.Name;
 import java.util.logging.Level;
