+++ src/main/java/net/kemitix/ldapmanager/navigation/events/SelectionChangeListener.java
@@ -22,9 +22,13 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package net.kemitix.ldapmanager.navigation;
+package net.kemitix.ldapmanager.navigation.events;
 
 import lombok.extern.java.Log;
+import net.kemitix.ldapmanager.navigation.NavigationItem;
+import net.kemitix.ldapmanager.navigation.events.NavigationItemOuSelectedEvent;
+import net.kemitix.ldapmanager.navigation.events.NavigationItemSelectionChangedEvent;
+import net.kemitix.ldapmanager.navigation.events.NavigationItemUserSelectedEvent;
 import net.kemitix.ldapmanager.state.LogMessages;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
