+++ src/test/java/net/kemitix/ldapmanager/navigation/events/NavigationItemSelectionChangedEventTest.java
@@ -1,5 +1,7 @@
-package net.kemitix.ldapmanager.navigation;
+package net.kemitix.ldapmanager.navigation.events;
 
+import net.kemitix.ldapmanager.navigation.NavigationItem;
+import net.kemitix.ldapmanager.navigation.events.NavigationItemSelectionChangedEvent;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
