+++ src/test/java/net/kemitix/ldapmanager/navigation/events/NavigationItemUserSelectedEventTest.java
@@ -1,6 +1,7 @@
-package net.kemitix.ldapmanager.navigation;
+package net.kemitix.ldapmanager.navigation.events;
 
 import net.kemitix.ldapmanager.domain.User;
+import net.kemitix.ldapmanager.navigation.UserNavigationItem;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
